requirements.txt
python-telegram-bot==20.6
notion-client==2.2.1
python-dotenv==1.0.1